Add block body test
